requirements.txt
# Core Framework - Railway Compatible
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# AI Providers
openai==1.3.7
anthropic==0.7.7

# ML and Data Science
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2

# Database and Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-multipart==0.0.6
aiosqlite==0.19.0
requests==2.31.0
aiofiles==23.2.1
jinja2==3.1.2

# Production server
gunicorn==21.2.0
